Fix error where time needed to be int
The make_chart file would not run because time in milliseconds was not an integer.

+++ settings.py
@@ -6,4 +6,4 @@ start_time = 2 # time to wait before program starts
 
 update_time = 0.1 # time between data queries
 
-update_time_ms = update_time * 100 # update_time in milliseconds
+update_time_ms = int(update_time * 100) # update_time in milliseconds
